Support Text Index Creation
Add support for users to create a text index. Each text
index contains an analyzer, a selector, and a list of fields.
The analyzer option allows the user to choose how Lucene
tokenizes the fields. The selector filters out unwanted documents
prior to search. The list of fields tells Lucene which
fields to index. If fields are not provided, all fields will be
indexed. Also, users can choose to enable a "default" field
that indexes all strings in the document.

33294-query-text-search