Support Text Index Creation
Add support for users to create a text index. Each text
index contains an analyzer, a selector, and a list of fields.
The analyzer contains option for Lucene to tokenize indexed fields.
The selector filters out documents prior to indexing.
The list of fields are used for Field:Val pairs for Lucene. Each
field also has a set of options for Lucene. Everything is returned
to dreyfus for text search indexing (powered by Lucene.)

33294-query-text-search