Introduce the $text operator to mango_selector
This just allows us to handle the $text operator in a selector. There is
a bit of a subtlety in that we apply the $text operator against a field
called $default which looks like an operator. We do this becaus the
query syntax for $text doesn't include a field which confuses a bit of
our other logic which always expects a field name to be present.

BugzId: 33294