Provide an option to disable the indexing of array lengths field.
Array length field is a single extra field per unique path to an array.
The case where we found this was a client that had data that used
arbitrary data as keys which exploded the number of fields in Lucene.
This is leading to JVM heap exhaustion as Lucene tries to cache
information about fields and Lucene is not designed to handle many
thousands fields.
The obvious fix was to switch to only indexing what they wanted to
query on. Unfortunately that didn't prevent the automatically created
array length fields from being created.
This patch is a big hammer to remove the auto generated array length
fields and we're also planning on another patch that removes array
length fields for anything that's not specified in the index's field
list.